fix: handle index error for updateBoard
Added a check to the updateBoard method inside the creatGameboard
factory function, to ensure that the passed index is greater than 0 and
smaller than the board array length. We want the board array to have
exactly 9 elements, so this is to ensure that no new elements can be
added, but existing elements can be changed.